fix: Add deterministic RSA key sorting for Windows timestamp resolution fix: Implement missing DatabaseProvider trait methods for RSA keypair management
fix: correct type safety for RSA keypair database persistence
Replace unsafe usize casts with i32 type for key_size_bits parameter

+++ src/database_plugins/postgres.rs
@@ -2663,7 +2663,7 @@ impl DatabaseProvider for PostgresDatabase {
         public_key_pem: &str,
         created_at: DateTime<Utc>,
         is_active: bool,
-        key_size_bits: usize,
+        key_size_bits: i32,
     ) -> Result<()> {
         sqlx::query(
             r"
@@ -2680,7 +2680,7 @@ impl DatabaseProvider for PostgresDatabase {
         .bind(public_key_pem)
         .bind(created_at)
         .bind(is_active)
-        .bind(key_size_bits as i32)
+        .bind(key_size_bits)
         .execute(&self.pool)
         .await?;
